refactor into more reusable components
